fix: new restructure of site closes #44 and #43.
Moved projects back to the front page as a component. blog will be a separate page with posts being written in markdown and saved to the db as such.

With the direct calls to db no need for the api routes to the blog & projects. Had to update the supabase client with NEXT_PUBLIC.